+++ src/main/java/com/example/test/controller/RideController.java
@@ -20,46 +20,46 @@ public class RideController {
     }
 
     @GetMapping("/active/{driverId}")
-    public Ride findDriversActiveRide(@PathVariable int id) {
+    public Ride findDriversActiveRide(@PathVariable Long id) {
         return service.findDriversActiveRide(id);
     }
 
     @GetMapping("/active/{passengerId}")
-    public Ride findPassengersActiveRide(@PathVariable int id) {
+    public Ride findPassengersActiveRide(@PathVariable Long id) {
         return service.findPassengersActiveRide(id);
     }
 
     @GetMapping("/{id}")
-    public Ride findRideById(@PathVariable int id) {
+    public Ride findRideById(@PathVariable Long id) {
         return service.findRideById(id);
     }
 
     // promijeniti status voznje
     @PutMapping("/{id}")
-    public boolean cancelExistingRide(@PathVariable int id) {
+    public boolean cancelExistingRide(@PathVariable Long id) {
         return service.cancelExistingRide(id);  //PRIJE NEGO VOZAC DODJE NA DESTINACIJU (to passenger moze)
     }
 
     //panic
     @PutMapping("/{id}/panic")
-    public Message setPanic(@RequestBody String reason, @PathVariable int id) {
+    public Message setPanic(@RequestBody String reason, @PathVariable Long id) {
         return service.setPanic(reason, id);
     }
 
     // prihvati voznju
     @PutMapping("/{id}/accept")
-    public Ride acceptRide(@PathVariable int id) {
+    public Ride acceptRide(@PathVariable Long id) {
         return service.acceptRide(id);
     }
 
     // prihvati voznju
     @PutMapping("/{id}/end")
-    public Ride endRide(@PathVariable int id) {
+    public Ride endRide(@PathVariable Long id) {
         return service.endRide(id);
     }
 
     @PutMapping("/{id}/cancel")
-    public Ride cancelRide(@RequestBody String reason, @PathVariable int id) {
+    public Ride cancelRide(@RequestBody String reason, @PathVariable Long id) {
         return service.cancelRide(reason, id);  //vozac moze da canceluje
     }
 }
